introduce internal delegate callback for jsi runtime initialization (#37605)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/37605

Changelog: [Internal]

in this stack, i remove the `bindingsInstallFunc` argument from `RCTHost`'s public API.

the bindingsInstallFunc is a lambda that provides a runtime parameter, which the consumer can "install bindings" to. we have our specific use case for this internally, but there's no need to expose this.

in this change, i introduce a delegate callback path that notifies an internal delegate that the react runtime has been created.

Reviewed By: javache

Differential Revision: D46215051

fbshipit-source-id: 732d6949064b0948b5cf883dc3c0ab494c2ecd94